test: Corrigido nome do arquivo e aumentado cobertura de teste
